Rename shutdownHandler -> shutdown
