+++ server/roll/generator.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 	"math"
-	//	"sr/shutdownHandler"
+	//	"sr/shutdown"
 )
 
 // Generator is an object which sends rolls from its RollSource through its channel.
